+++ R/RcppExports.R
@@ -28,7 +28,7 @@
 #'
 #'
 #' \code{adjusted_asymmetric_accuracy()} (Gagolewski, 2022)
-#' only accepts \eqn{K = L}. It is an external cluster validity measure
+#' is an external cluster validity measure
 #' which assumes that the label vector \code{x} (or rows in the confusion
 #' matrix) represents the reference (ground truth) partition.
 #' It is a corrected-for-chance summary of the proportion of correctly
@@ -39,15 +39,15 @@
 #' where \eqn{C} is the confusion matrix.
 #'
 #' \code{normalized_accuracy()} is defined as
-#' \eqn{(Accuracy(C_\sigma)-1/L)/(1-1/L)}, where \eqn{C_\sigma} is a version
-#' of the confusion matrix for given \code{x} and \code{y},
-#' \eqn{K \leq L}, with columns permuted based on the solution to the
+#' \eqn{(Accuracy(C_\sigma)-1/max(K,L))/(1-1/max(K,L))}, where \eqn{C_\sigma} is a version
+#' of the confusion matrix for given \code{x} and \code{y}
+#' with columns permuted based on the solution to the
 #' maximal linear sum assignment problem.
 #' The \eqn{Accuracy(C_\sigma)} part is sometimes referred to as
 #' set-matching classification rate or pivoted accuracy.
 #'
 #' \code{pair_sets_index()} gives the Pair Sets Index (PSI)
-#' adjusted for chance (Rezaei, Franti, 2016), \eqn{K \leq L}.
+#' adjusted for chance (Rezaei, Franti, 2016).
 #' Pairing is based on the solution to the linear sum assignment problem
 #' of a transformed version of the confusion matrix.
 #' Its simplified version assumes E=1 in the definition of the index,